requirements.txt
# Common
orjson==3.9.15
pydantic==2.8.2
pydantic-settings==2.2.1
python-cqrs==3.0.1
python-dotenv==1.0.0

# API
fastapi_app@git+https://github.com/vadikko2/fastapi-app@0.0.2
redis==5.0.3
fastapi==0.109.2

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Images
blurhash==1.1.5
simplejpeg==1.8.2
pillow==10.4.0

# S3
boto3==1.34.131

# Mock
faker==36.1.1
faker_music==0.4

# Http
httpx==0.27.0
httpx-retries==0.4.0

# Auth
pyjwt[crypto]==2.10.1
